handled exception

+++ backend/src/main/java/com/htecgroup/skynest/service/impl/PermissionServiceImpl.java
@@ -6,6 +6,7 @@ import com.htecgroup.skynest.exception.buckets.BucketAccessDeniedException;
 import com.htecgroup.skynest.exception.buckets.BucketAlreadyDeletedException;
 import com.htecgroup.skynest.exception.buckets.BucketNotFoundException;
 import com.htecgroup.skynest.exception.permission.PermissionAlreadyExistsException;
+import com.htecgroup.skynest.exception.permission.PermissionDoesNotExistException;
 import com.htecgroup.skynest.model.entity.*;
 import com.htecgroup.skynest.model.request.PermissionEditRequest;
 import com.htecgroup.skynest.model.request.PermissionGrantRequest;
@@ -184,7 +185,14 @@ public class PermissionServiceImpl implements PermissionService {
     UserEntity user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
     UserObjectAccessEntity permission =
         permissionRepository.findByObjectIdAndGrantedTo(bucketId, user);
+    checkIfPermissionExist(permission);
     permissionRepository.delete(permission);
+  }
+
+  private void checkIfPermissionExist(UserObjectAccessEntity permission) {
+    if (permission == null) {
+      throw new PermissionDoesNotExistException();
+    }
   }
 
   private UserEntity findTargetUserForEdit(PermissionEditRequest permissionEditRequest) {
